requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiostream==0.6.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astra-toolbox==2.2.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.15
docarray==0.40.0
docker==7.1.0
einops==0.8.0
einops-exts==0.0.4
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
gdown==5.2.0
grpcio==1.69.0
grpcio-health-checking==1.48.2
grpcio-reflection==1.48.2
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.27.1
humanize==4.11.0
idna==3.4
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.3
lazy_loader==0.4
lightning-utilities==0.11.9
lz4==3.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl-service==2.4.1
mkl_fft==1.3.8
mkl_random==1.2.4
monai==1.4.0
monai-generative==0.2.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.2.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==3.20.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydicom==3.0.1
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.28.1
rich==13.9.4
rotary-embedding-torch==0.8.6
safetensors==0.5.2
scikit-image==0.25.0
scikit-video==1.1.11
scipy==1.15.1
shellingham==1.5.4
SimpleITK==2.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tifffile==2025.1.10
timm==1.0.14
torchio==0.20.3
torchmetrics==1.6.1
tqdm==4.67.1
triton==2.2.0
typer==0.15.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.13
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.1
yarl==1.18.3